pyproject.toml
[project]
name = "civicrm-analyse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ansible>=11.8.0",
    "ansible-lint>=25.7.0",
    "infisicalsdk>=1.0.9",
    "jupyter>=1.1.1",
    "polars>=1.32.0",
    "requests>=2.32.4",
    "python_tools",
    "python-dotenv>=1.1.1",
    "faker>=37.5.3",
]

[tool.uv.sources]
python_tools = { workspace = true }

[tool.uv.workspace]
members = ["python-tools"]
